+++ src/cxx/bridge/mod.rs
@@ -1,1 +1,3 @@
+pub mod fingerprint;
+pub mod mol_standardize;
 pub mod rdmol;
